Use url for hashable/equatable (IOS-190)
and do some refactoring
