+++ MastodonSDK/Sources/MastodonCore/Extension/MastodonSDK/Mastodon+Entity+Account.swift
@@ -9,33 +9,6 @@ import Foundation
 import MastodonSDK
 import MastodonMeta
 
-extension Mastodon.Entity.Account: Hashable {
-    public func hash(into hasher: inout Hasher) {
-        hasher.combine(id)
-    }
-    
-    public static func == (lhs: Mastodon.Entity.Account, rhs: Mastodon.Entity.Account) -> Bool {
-        return lhs.id == rhs.id
-    }
-}
-
-extension Mastodon.Entity.Account {    
-    public func avatarImageURL() -> URL? {
-        let string = UserDefaults.shared.preferredStaticAvatar ? avatarStatic ?? avatar : avatar
-        return URL(string: string)
-    }
-    
-    public func avatarImageURLWithFallback(domain: String) -> URL {
-        return avatarImageURL() ?? URL(string: "https://\(domain)/avatars/original/missing.png")!
-    }
-}
-
-extension Mastodon.Entity.Account {
-    public var displayNameWithFallback: String {
-        return !displayName.isEmpty ? displayName : username
-    }
-}
-
 extension Mastodon.Entity.Account {
     public var emojiMeta: MastodonContent.Emojis {
         let isAnimated = !UserDefaults.shared.preferredStaticEmoji
